+++ src/core/http.ts
@@ -1,29 +1,25 @@
-import {
-  RequestPayload,
-  RequestOptions,
-  KlepperResponse,
-  KlepperIncomingMessage,
-  RequestStatus,
-} from "../transport";
 import * as http from "http";
+import { RequestStatus } from "../transport/enums";
+import { EventResponse, KlepperEvent } from "../transport/events";
+import { KlepperIncomingMessage, RequestOptions } from "../transport/http";
 import { getGlobalClientData } from "./global";
+import { isClientConnected } from "./is";
 
 const KLEPPER_HOST = process.env.KLEPPER_HOST || "localhost";
 const KLEPPER_API = process.env.KLEPPER_API || "/test";
 const KLEPPER_PORT = process.env.KLEPPER_PORT || 3000;
 
-const createHttpOptions = (payload: RequestPayload): http.RequestOptions => {
+const createHttpOptions = (event: KlepperEvent): http.RequestOptions => {
   const client = getGlobalClientData();
 
   const { privateKey } = client;
-  const { data } = payload;
   const baseOptions: RequestOptions = {
     hostname: KLEPPER_HOST,
     port: +KLEPPER_PORT,
     method: "POST",
     headers: {
       "Content-Type": "application/json",
-      "Content-Length": `${Buffer.byteLength(JSON.stringify(data))}`,
+      "Content-Length": `${Buffer.byteLength(JSON.stringify(event))}`,
       "Klepper-Project-Key": privateKey as string,
     },
   };
@@ -37,60 +33,60 @@ const createHttpOptions = (payload: RequestPayload): http.RequestOptions => {
 const statusFromCode = (code: number) =>
   code >= 200 && code <= 299 ? RequestStatus.SUCCESS : RequestStatus.ERROR;
 
-/**
- * Send data to Klepper server
- *
- * @param payload
- */
 export const sendEvent = async (
-  payload: RequestPayload
-): Promise<KlepperResponse> => {
-  const { data } = payload;
-
-  try {
-    return new Promise<KlepperResponse>((resolve, reject) => {
-      const httpOptions = createHttpOptions(payload);
-      const client = getGlobalClientData();
-
-      if (!httpOptions) {
-        reject(new Error("NO HTTP OPTIONS"));
-      }
-
-      if (!client) {
-        reject(new Error("NO CLIENT DATA IN GLOBAL OBJECT"));
-      }
-
-      const request = http.request(
-        httpOptions,
-        (res: KlepperIncomingMessage) => {
-          res.setEncoding("utf8");
-
-          const status = statusFromCode(res?.statusCode as number);
-          const isSuccess = status === RequestStatus.SUCCESS;
-
-          if (!isSuccess) {
-            reject(new Error("HTTP ERROR: ${res.statusCode}"));
-          } else {
-            resolve({
-              statusCode: res?.statusCode as number,
-              statusMessage: "Event successfully sended to Klepper",
-            });
-          }
-
-          res.on("error", reject);
-        }
-      );
-      request.on("error", reject);
-
-      request.on("timeout", () => {
-        request.destroy();
-        reject(new Error("TIMEOUT"));
+  event: KlepperEvent
+): Promise<EventResponse> => {
+  return new Promise<EventResponse>((resolve, reject) => {
+    const httpOptions = createHttpOptions(event);
+    if (!httpOptions) {
+      reject({
+        statusCode: 400,
+        statusMessage:
+          "[Klepper] Error during sending event to Klepper. No HTTP options.",
       });
+    }
 
-      request.write(JSON.stringify(data));
-      request.end();
+    if (!isClientConnected()) {
+      reject({
+        statusCode: 400,
+        statusMessage:
+          "[Klepper] Error during sending event to Klepper. No client global data in NodeJS scope.",
+      });
+    }
+
+    const request = http.request(httpOptions, (res: KlepperIncomingMessage) => {
+      res.setEncoding("utf8");
+
+      const status = statusFromCode(res?.statusCode as number);
+      const isSuccess = status === RequestStatus.SUCCESS;
+
+      if (!isSuccess) {
+        reject({
+          statusCode: res?.statusCode as number,
+          statusMessage: "[KLEPPER] Error during sending event to Klepper.",
+        });
+      } else {
+        resolve({
+          statusCode: res?.statusCode as number,
+          statusMessage: "[KLEPPER] Event successfully sended to Klepper.",
+        });
+      }
+
+      res.on("error", reject);
     });
-  } catch (error) {
-    throw new Error(""); //to properly handling
-  }
+    
+    request.on("error", reject);
+
+    request.on("timeout", () => {
+      request.destroy();
+      reject({
+        statusCode: 400,
+        statusMessage:
+          "[Klepper] Error during sending event to Klepper. Connection timeout.",
+      });
+    });
+
+    request.write(JSON.stringify(event));
+    request.end();
+  });
 };