+++ src/core/index.ts
@@ -1,4 +1,3 @@
-import * as http from "./http";
 import * as global from "./global";
 
-export { http, global };
+export { global };