+++ src/core/types/global.d.ts
@@ -1,4 +1,4 @@
-import { KlepperGlobal } from "../../transport";
+import { KlepperGlobal } from "../../transport/base";
 
 declare global {
   var __KLEPPER__: KlepperGlobal;