+++ src/node/index.ts
@@ -1,4 +1,4 @@
-import * as Middleware from "./middlewares/index";
+import { Middleware } from "..";
 import * as Klepper from "./sdk";
 
 export { Middleware, Klepper };