+++ src/node/middlewares/index.ts
@@ -1,3 +1,0 @@
-import { errorMiddleware } from "./error";
-
-export { errorMiddleware as ErrorMiddleware };