+++ src/node/sdk.ts
@@ -1,8 +1,8 @@
-import { global } from "../core";
-import { KlepperOptions } from "../transport";
-import { helpers } from "./helpers";
+import { setGlobalClientData } from "../core/global";
+import { isClientConnected } from "../core/is";
+import { KlepperOptions } from "../transport/options";
 
-const defaultConnectCallback = () => true;
+export const defaultBooleanCallback = () => true;
 
 /**
  *
@@ -12,25 +12,33 @@ const defaultConnectCallback = () => true;
  * @param callback default return true, if return false then connection to Klepper has not been established
  *
  * @example
- * Klepper.connect({ credentials: { apiKey: "", secretKey: "" } }, () => {
+ *
+ * ```
+ * Klepper.connect({ credentials: { privateKey: "" } }, () => {
  *   if (process.env.ENV === "prod") {
  *      return true;
  *  }
  *   return false;
  * })
+ * ```
  */
 export const connect = (
   options: KlepperOptions,
-  callback = defaultConnectCallback
+  callback = defaultBooleanCallback
 ): void => {
-  const client = global.getGlobalClientData();
+  /**
+   * TODO:
+   *
+   * check if @options > @projectId is existing in Klepper db
+   */
+
   if (!callback) {
     console.warn(
       `[KLEPPER]: Due to a callback setting for the connect function, the attempt to establish a connection has failed.`
     );
   }
 
-  if (helpers.isEmpty(client)) {
-    global.setGlobalClientData({ privateKey: options.privateKey });
+  if (!isClientConnected()) {
+    setGlobalClientData({ privateKey: options.privateKey });
   }
 };