+++ src/transport/options.ts
@@ -1,0 +1,17 @@
+import { ExceptionPriority } from "./enums";
+import { Environment } from "./types";
+
+export interface KlepperOptions {
+  environment?: Environment;
+  privateKey: string;
+}
+
+export interface ErrorMiddlewareOptions {
+  allowLocalhost?: boolean;
+  allowHttp?: boolean;
+}
+
+export interface CatchExceptionsOptions {
+  tag?: string;
+  priority?: ExceptionPriority;
+}