+++ src/transport/types.ts
@@ -1,0 +1,3 @@
+export type RequestMethodType = "GET" | "POST" | "PUT" | "PATCH" | "DELETE";
+
+export type Environment = undefined | "prod" | "dev" | "test";