+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # CHANGELOG
 
+## 0.0.3-alpha - 2022-03-23
+- Split transport file into few small files
+- removing try/catch from various functions
+- added options in parameters to `catchException` function
+- fixes for middleware and process of handling exceptions
+- fix for build script
 ## 0.0.2-alpha - 2022-03-23
 - Migration from monorepo in lerna to single monolith
 - Bugfixes